SWORD25: Resolve ambigous write() call on 64 bit systems
svn-id: r53410

+++ engines/sword25/kernel/outputpersistenceblock.cpp
@@ -108,7 +108,7 @@ void OutputPersistenceBlock::write(const Common::String &String) {
 void OutputPersistenceBlock::write(const void *BufferPtr, size_t Size) {
 	WriteMarker(BLOCK_MARKER);
 
-	write(Size);
+	write((int) Size);
 	RawWrite(BufferPtr, Size);
 }
